refactor: extract login interfaces into login.d.ts
